requirements.txt
# ===== FastAPI Core =====
fastapi==0.109.2
uvicorn==0.29.0
python-dotenv
python-multipart
passlib[bcrypt]

# ===== Audio Processing =====
numpy==1.26.4
librosa
soundfile
scipy
tqdm

# ===== Voice Embedding =====
speechbrain==1.0.3
huggingface-hub==0.19.4

# ===== AI / LLM =====
google-generativeai==0.5.4

# ===== Google APIs (optional, safe to keep) =====
google-api-python-client
google-auth
google-auth-httplib2

# ===== PyTorch CPU ONLY =====
torch==2.1.2+cpu
torchaudio==2.1.2+cpu
